a snapshot of experimental parameter settings
	modified:   src/config.rs
	modified:   src/processor/mod.rs
	modified:   src/solver/conflict.rs
	modified:   src/solver/restart.rs
	modified:   src/solver/search.rs
	modified:   src/state.rs

+++ src/solver/search.rs
@@ -254,8 +254,8 @@ fn search(
         }
         // Simplification has been postponed because chronoBT was used.
         if asg.decision_level() == asg.root_level {
-            if use_vivify && state.config.viv_int <= state.to_vivify {
-                state.to_vivify = 0;
+            if use_vivify && state.config.viv_int <= state.to_vivify as usize {
+                state.to_vivify = 0.0;
                 if vivify(asg, cdb, elim, state).is_err() {
                     // return Err(SolverError::UndescribedError);
                     analyze_final(asg, state, &cdb[ci]);
